#78
ContoCorrenteTest NullPointerException, because the return value of "it.iad.sigeba2.dto.ContoCorrenteDto.getId()" is null

da Sistemare

+++ server/src/main/java/it/iad/sigeba2/service/impl/AnagraficaClientiServiceImpl.java
@@ -6,6 +6,9 @@ import it.iad.sigeba2.dto.CriterioClienteDto;
 import it.iad.sigeba2.dto.CriterioInserimentoClienteDto;
 import it.iad.sigeba2.dto.CriterioModificaClienteDto;
 import it.iad.sigeba2.dto.SimpleIdDto;
+import it.iad.sigeba2.enumerated.GravitaStatoEnum;
+import it.iad.sigeba2.exception.SigebaException;
+import it.iad.sigeba2.helper.SigebaStateCollector;
 import it.iad.sigeba2.model.Cliente;
 import it.iad.sigeba2.repository.ClienteRepository;
 import it.iad.sigeba2.service.AnagraficaClientiService;
@@ -23,11 +26,16 @@ public class AnagraficaClientiServiceImpl implements AnagraficaClientiService {
     private ClienteRepository clienteRepository;
 
     @Override
-    public List<ClienteDto> cercaCliente(CriterioClienteDto criterio) {
+    public List<ClienteDto> cercaCliente(CriterioClienteDto criterio) throws SigebaException {
 
         log.debug("Entrato in cercaCliente");
+        if (criterio == null || criterio.getCriterio() == null) {
+            SigebaStateCollector.addStatusMessage("Criterio di ricerca null non consentito",
+                    "Inserire un criterio di ricerca valido", GravitaStatoEnum.CRITICA);
+            throw new SigebaException("Criterio di ricerca null non ammissibile");
+        }
         String crit = criterio.getCriterio();
-        List<ClienteDto> dtos = new ArrayList<>();
+        List<ClienteDto> listaClienti = new ArrayList<>();
 
         // trova i clienti
         String chiaveParziale = "%" + crit + "%";
@@ -35,54 +43,73 @@ public class AnagraficaClientiServiceImpl implements AnagraficaClientiService {
 
         // li converte in lista di DTO
         for (Cliente cliente : clienti) {
-            dtos.add(new ClienteDto(cliente));
+            listaClienti.add(new ClienteDto(cliente));
         }
         log.debug("Uscito da cercaCliente");
-        return dtos;
+        return listaClienti;
     }
 
     @Override
-    public List<ClienteDto> inserisciCliente(CriterioInserimentoClienteDto dto) {
+    public List<ClienteDto> inserisciCliente(CriterioInserimentoClienteDto clienteDaInserireDto) throws SigebaException {
         log.debug("Entra in inserisciCliente");
-        // riceve il DTO e lo trasforma in Cliente
-        Cliente cliente = new Cliente(dto.getCliente());
+        if (clienteDaInserireDto == null || clienteDaInserireDto.getCliente() == null) {
+            SigebaStateCollector.addStatusMessage("Cliente null non consentito",
+                    "Inserire un cliente valido", GravitaStatoEnum.CRITICA);
+            throw new SigebaException("Cliente da inserire null!!!");
+        }
 
-        // aggiunge il dto alla lista dei clienti
+        Cliente cliente = new Cliente(clienteDaInserireDto.getCliente());
+
         clienteRepository.save(cliente);
 
-        // chiama il metodo cercaCliente per ritornare la lista filtrata
         log.debug("Esci da inserisciCliente");
-        List<ClienteDto> clienteDto = cercaCliente(dto.getFiltro());
+        List<ClienteDto> clienteDto = cercaCliente(clienteDaInserireDto.getFiltro());
         return clienteDto;
     }
 
     @Override
-    public Cliente leggiCliente(SimpleIdDto chiave) {
+    public Cliente leggiCliente(SimpleIdDto chiave) throws SigebaException {
+        log.debug("Entrato in modifica cliente");
+        if (chiave == null || chiave.getId() == null) {
+            SigebaStateCollector.addStatusMessage("Id da leggere null non consentito",
+                    "Inserire Id cliente valido", GravitaStatoEnum.CRITICA);
+            throw new SigebaException("Id cliente da leggere null!!");
+        }
         return clienteRepository.findById(chiave.getId())
                 .map(cx -> cx)
                 .orElse(null);
     }
 
     @Override
-    public List<ClienteDto> modificaCliente(CriterioModificaClienteDto modificaDto) {
+    public List<ClienteDto> modificaCliente(CriterioModificaClienteDto modifica) throws SigebaException {
         log.debug("Entrato in modificaClienti");
+        if (modifica == null || modifica.getCliente() != null) {
+            SigebaStateCollector.addStatusMessage("Cliente da modificare null non consentito",
+                    "Inserirre cliente valido", GravitaStatoEnum.CRITICA);
+            throw new SigebaException("Cliente da modificare null!!!");
+        }
 
-        ClienteDto clienteModificato = modificaDto.getCliente();
+        ClienteDto clienteModificato = modifica.getCliente();
         Cliente clienteCheSostituisce = new Cliente(clienteModificato);
 
         clienteRepository.save(clienteCheSostituisce);
 
         log.debug("Esce da modificaClienti");
-        return cercaCliente(modificaDto.getFiltro());
+        return cercaCliente(modifica.getFiltro());
     }
 
     @Override
-    public List<ClienteDto> cancellaCliente(CriterioCancellazioneClienteDto dtoCancellazione) {
+    public List<ClienteDto> cancellaCliente(CriterioCancellazioneClienteDto dtoCancellazione) throws SigebaException {
+
         log.debug("Entrato in cancellaCliente");
-        // recupera l'id del cliente da rimuovere
+        if (dtoCancellazione == null || dtoCancellazione.getIdCliente() == null) {
+            SigebaStateCollector.addStatusMessage("Id da cancellare null non consentito",
+                    "Inserire un Id valido", GravitaStatoEnum.CRITICA);
+            throw new SigebaException("Id da cancellare null!!");
+        }
+
         Long idDaRimuovere = dtoCancellazione.getIdCliente();
 
-        // lo rimuove
         clienteRepository.deleteById(idDaRimuovere);
 
         // recupera i clienti rimasti
